fix(released): upgrade download method sftp.

+++ sample/WeNet/scripts/download.sh
@@ -1,34 +1,36 @@
 #!/bin/bash
-pip3 install dfn
-# sudo apt install unzip
-
+res=$(dpkg -l|grep unzip)
+if [ $? != 0 ];
+then
+    echo "Please install unzip on your system!"
+    exit
+fi
+pip3 install dfss -i https://pypi.tuna.tsinghua.edu.cn/simple --upgrade
 scripts_dir=$(dirname $(readlink -f "$0"))
-# echo $scripts_dir
- 
+
 pushd $scripts_dir
 # datasets
 if [ ! -d "../datasets" ]; 
 then
     mkdir -p ../datasets
-    # aishell data
-    python3 -m dfn --url http://219.142.246.77:65000/sharing/JigSE5PuB
-    unzip aishell_S0764.zip -d ../datasets/
+    python3 -m dfss --url=open@sophgo.com:sophon-demo/WeNet/datasets/aishell_S0764.zip
+    unzip aishell_S0764.zip -d ../
     rm aishell_S0764.zip
- 
+
     echo "datasets download!"
 else
-    echo "datasets exist!"
+    echo "Datasets folder exist! Remove it if you need to update."
 fi
- 
+
 # models
 if [ ! -d "../models" ]; 
 then
-    python3 -m dfn --url http://219.142.246.77:65000/sharing/COB3NbyUj
+    python3 -m dfss --url=open@sophgo.com:sophon-demo/WeNet/models_0918/models.zip
     unzip models.zip -d ../
     rm models.zip
     echo "models download!"
 else
-    echo "models exist!"
+    echo "Models folder exist! Remove it if you need to update."
 fi
 
 # swig decoders module
@@ -36,13 +38,13 @@ if [ ! -d "../python/swig_decoders" ];
 then
     if [ "$1" = "soc" ];
     then
-        python3 -m dfn --url http://219.142.246.77:65000/sharing/FBii7gQRZ
+        python3 -m dfss --url=open@sophgo.com:sophon-demo/WeNet/soc/swig_decoders_aarch64.zip
         unzip swig_decoders_aarch64.zip -d ../python/
         mv ../python/swig_decoders_aarch64 ../python/swig_decoders
         rm swig_decoders_aarch64.zip
     elif [ "$1" = "pcie" ];
     then
-        python3 -m dfn --url http://219.142.246.77:65000/sharing/N6GDGWyOg
+        python3 -m dfss --url=open@sophgo.com:sophon-demo/WeNet/pcie/swig_decoders_x86_64.zip
         unzip swig_decoders_x86_64.zip -d ../python/
         mv ../python/swig_decoders_x86_64 ../python/swig_decoders
         rm swig_decoders_x86_64.zip
